[frontend] WIP container add stix core object enforce reference handling
