+++ opencti-platform/opencti-front/src/private/components/common/containers/ContainerAddStixCoreObjects.jsx
@@ -70,6 +70,7 @@ const ContainerAddStixCoreObjects = (props) => {
     selectedText,
     openDrawer,
     handleClose,
+    enableReferences,
   } = props;
   const classes = useStyles();
   const { t_i18n } = useFormatter();
@@ -368,6 +369,7 @@ const ContainerAddStixCoreObjects = (props) => {
               setNumberOfElements={helpers.handleSetNumberOfElements}
               mapping={mapping}
               containerRef={containerRef}
+              enableReferences={enableReferences}
             />
           )}
         />
